Initial release.
dpprint - for date prefixed print - is a simple utility which reads on
stdin and prefixes every line with the current date.
It is useful to know when a message has been printed by a program which
takes long to run.
Note that dpprint is relatively useless, as it is a very simple wrapper
on gawk(1).